backend/requirements/development.txt
Django>=2.1.3,<2.2.0
djangorestframework>=3.9.0,<3.10.0
python-interface
redis>=3.3.11,<3.3.12
django-environ
tenacity>=5.1.1,<5.1.2
pytest>=5.2.2,<5.2.3
shortuuid>=0.5.0,<0.5.1
random-word>=1.0.4,<1.0.5
grpcio-tools==1.24.3

flake8>=3.6.0,<3.7.0
autopep8>1.4.0,<=1.4.4
python-coveralls
coverage
black